requirements.txt
websockets
python-telegram-bot==21.4
python-dotenv